requirements.txt
boto3==1.38.38
python-dotenv==1.1.1
websockets==15.0.1
Requests==2.32.4
pandas==2.3.1
pyarrow==21.0.0